+++ spng/src/lib.rs
@@ -391,3 +391,8 @@ where
     reader.next_frame(&mut out)?;
     Ok((out_info, out))
 }
+
+/// Returns the `libspng` version: `(major, minor, patch)`
+pub fn version() -> (u32, u32, u32) {
+    (spng_sys::SPNG_VERSION_MAJOR, spng_sys::SPNG_VERSION_MINOR, spng_sys::SPNG_VERSION_PATCH)
+}