+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0-alpha.1] - 2021-06-23
 ### Added
 - Update to libspng `0.7.0-rc2`
 - `Limits`, `Info`, and `OutputInfo` now implment `Debug`, `Copy`, `Clone`, `PartialEq`, `Eq`, `Hash`
 - `RawContext::get_hist`
 - `RawContext::get_unknown_chunks`
-- The major, minor, and patch version can be queried with `spng::version()`
+- The libspng `major`, `minor`, and `patch` version can be queried with `spng::version()`
 
 ## [0.1.0] - 2021-03-14
 ### Changed
@@ -55,7 +57,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Initial rust wrapper with minimal API surface
 - Initial native bindings to [libspng] `master` ([2079ef6])
 
-[Unreleased]: https://github.com/aloucks/spng-rs/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/aloucks/spng-rs/compare/v0.2.0-alpha.1...HEAD
+[0.2.0-alpha.1]: https://github.com/aloucks/spng-rs/releases/tag/v0.2.0-alpha.1
 [0.1.0]: https://github.com/aloucks/spng-rs/releases/tag/v0.1.0
 [0.1.0-alpha.6]: https://github.com/aloucks/spng-rs/releases/tag/v0.1.0-alpha.6
 [0.1.0-alpha.5]: https://github.com/aloucks/spng-rs/releases/tag/v0.1.0-alpha.5